int32

+++ atm_dyn_iconam/src/icon4py/atm_dyn_iconam/mo_solve_nonhydro_stencil_52.py
@@ -12,14 +12,14 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from gt4py.next.ffront.decorator import field_operator, program, scan_operator
-from gt4py.next.ffront.fbuiltins import Field
+from gt4py.next.ffront.fbuiltins import Field, int32
 
 from icon4py.common.dimension import CellDim, KDim, Koff
 
 
-@scan_operator(axis=KDim, forward=True, init=(0.0, 0.0, 1))
+@scan_operator(axis=KDim, forward=True, init=(0.0, 0.0, int32(1)))
 def _w(
-    state: tuple[float, float, int],
+    state: tuple[float, float, int32],
     w: float,
     z_q: float,
     z_a: float,
@@ -31,7 +31,7 @@ def _w(
     z_g = 1.0 / (z_b + z_a * z_q_m1)
     z_q_new = (0.0 - z_c) * z_g
     w_new = (w_prep - z_a * w_m1) * z_g
-    return (z_q, w, 0) if first == 1 else (z_q_new, w_new, 0)
+    return (z_q, w, int32(0)) if first == int32(1) else (z_q_new, w_new, int32(0))
 
 
 @field_operator
@@ -50,7 +50,7 @@ def _mo_solve_nonhydro_stencil_52(
 ) -> tuple[
     Field[[CellDim, KDim], float],
     Field[[CellDim, KDim], float],
-    Field[[CellDim, KDim], int],
+    Field[[CellDim, KDim], int32],
 ]:
     z_gamma = dtime * cpd * vwind_impl_wgt * theta_v_ic / ddqz_z_half
     z_a = (0.0 - z_gamma) * z_beta(Koff[-1]) * z_alpha(Koff[-1])
@@ -120,7 +120,7 @@ def mo_solve_nonhydro_stencil_52(
     w: Field[[CellDim, KDim], float],
     dtime: float,
     cpd: float,
-    dummy_bool: Field[[CellDim, KDim], int],
+    dummy_bool: Field[[CellDim, KDim], int32],
 ):
     _mo_solve_nonhydro_stencil_52(
         vwind_impl_wgt,
